requirements.txt
coverage
coveralls
flask
flask-login
flask-sqlalchemy
flask-testing
mock
mysql-python
pbr
pre-commit
pycrypto
python-ldap
requests
six
